// WebUI/src/app/core/components/app/product-enter/models/productEnterDto.ts
export class ProductEnterDto{
    productCode: string;
    productName: string;
    modelId: number;
    model: string;
    unitTypeId: number;
    unitType: string;
    shelfPlaceId: number;
    shelfPlace: string;
    firmId: number;
    firm: string;
    purchasePrice: number;
    salePrice: number;
    description: string;
}